ai-service/requirements.txt
# fastapi==0.104.1
# uvicorn==0.24.0
# pandas==2.1.1
# numpy==1.24.3
# python-dotenv==1.0.0

fastapi==0.110.0
uvicorn==0.27.1
pandas==2.2.1
numpy==1.26.4
python-dotenv==1.0.1
